show alert to notify user the function has finished

+++ Lessions/Lession_7/AngularJS_Study/public/angular-app/src/app/game/game.component.ts
@@ -35,17 +35,20 @@ export class GameComponent implements OnInit {
   }
 
   onDelete(game: Game) {
-    this._gameService.deleteOne(game._id).subscribe({
-      next: () => {
-        this._router.navigate(["games"]);
-      },
-      error: (error) => {
-        console.log(error);
-      },
-      complete: () => {
-        console.log("Completed!");
-      }
-    });
+    if (confirm("Do you want to delete " + game.title + "?")) {
+      this._gameService.deleteOne(game._id).subscribe({
+        next: () => {
+          this._router.navigate(["games"]);
+        },
+        error: (error) => {
+          console.log(error);
+        },
+        complete: () => {
+          console.log("Completed!");
+        }
+      });
+    }
+    
   }
 
 }
